+++ pyfakefs/extra_packages.py
@@ -16,13 +16,8 @@ If the external module is not present, the build-in module is imported.
 
 try:
     import pathlib2
-
-    pathlib = pathlib2
 except ImportError:
     pathlib2 = None
-    import pathlib
-
-    pathlib = pathlib
 
 try:
     import scandir