+++ pyfakefs/tests/all_tests_without_extra_packages.py
@@ -11,20 +11,12 @@
 # limitations under the License.
 
 """A test suite that runs all tests for pyfakefs at once.
-Excludes tests using external pathlib2 and scandir packages."""
+Excludes tests using external scandir package."""
 
 import sys
 import unittest
 
 from pyfakefs import extra_packages
-
-if extra_packages.pathlib2:
-    extra_packages.pathlib2 = None
-    try:
-        import pathlib
-    except ImportError:
-        pathlib = None
-    extra_packages.pathlib = pathlib
 
 if extra_packages.use_scandir_package:
     extra_packages.use_scandir_package = False