+++ CHANGES.md
@@ -1,7 +1,10 @@
 # pyfakefs Release Notes
 The released versions correspond to PyPi releases.
 
-## Version 4.5.0 (as yet unreleased)
+## Version 4.6.0 (as yet unreleased)
+
+## [Version 4.5.0](https://pypi.python.org/pypi/pyfakefs/4.5.0) (2021-06-04)
+Adds some support for Python 3.10 and basic type checking.
 
 ### New Features
   * added support for some Python 3.10 features: